src/cores/cicada-ingestion/data-providers/Github/sources/ingestion-github/src/requirements.txt
pandas
protobuf
PyGithub
python-dotenv
Requests
schedule
tqdm
urllib3
github
pyarrow
